# packages/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/basics/inspectors": {
            "get": {
                "description": "This endpoint retrieves ticket inspector reports from the database within a specified time range.\nIt supports filtering by start and end timestamps,\nand checks if the data has been modified since the last request using the \"If-Modified-Since\" header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "basics"
                ],
                "summary": "Retrieve information about ticket inspector reports",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start timestamp (RFC3339 format)",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End timestamp (RFC3339 format)",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/utils.TicketInspectorResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Accepts a JSON payload with details about a ticket inspector's current location.\nThis endpoint validates the provided data, processes necessary computations for linking stations and lines,\ninserts the data into the database, and triggers an update to the risk model used in operational analysis.\nIf the 'timestamp' field is not provided in the request, the current UTC time truncated to the nearest minute is used automatically.\nThe endpoint also includes a rate limit to prevent abuse. The rate limit is based on the IP address of the request.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "basics"
                ],
                "summary": "Submit ticket inspector data",
                "parameters": [
                    {
                        "description": "Data about the inspector's location and activity",
                        "name": "inspectorData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.InspectorRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Timestamp of the report in ISO 8601 format (e.g., 2006-01-02T15:04:05Z); if not provided, the current time is used",
                        "name": "timestamp",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully processed and inserted the inspector data with computed linkages and risk model updates.",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseData"
                        }
                    },
                    "400": {
                        "description": "Bad Request: Missing or incorrect parameters provided."
                    },
                    "429": {
                        "description": "Too Many Requests: The request has been rate limited."
                    },
                    "500": {
                        "description": "Internal Server Error: Error during data processing or database insertion."
                    }
                }
            }
        },
...
        "/stations/{stationId}/statistics": {
            "get": {
                "description": "Retrieves statistics for a specific station.\nThis endpoint returns the number of reports for the specified station within the given time range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stations"
                ],
                "summary": "Get station statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the station",
                        "name": "stationId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID of the line (optional)",
                        "name": "lineId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time for the statistics (format: RFC3339)",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time for the statistics (format: RFC3339)",
                        "name": "end",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved station statistics.",
                        "schema": {
                            "$ref": "#/definitions/statistics.Statistics"
                        }
                    },
                    "400": {
                        "description": "Bad Request: Invalid time format.",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error: Failed to get number of reports.",
                        "schema": {}
                    }
                }
            }
        },
...
        "/v0/transit/itineraries": {
            "get": {
                "description": "Retrieves possible routes between two stations, including a safest itinerary based on risk prediction.\nThis endpoint calculates multiple itinerary options and enriches them with risk data from the risk prediction model.\nThe response includes both the safest itinerary and alternative itineraries, sorted by their calculated risk.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transit"
                ],
                "summary": "Get route itineraries between two stations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start station ID",
                        "name": "startStation",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End station ID",
                        "name": "endStation",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved route options",
                        "schema": {
                            "$ref": "#/definitions/itineraries.ItinerariesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request: Missing or invalid station IDs",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error: Failed to process route data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway: Failed to fetch route from engine",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "itineraries.ItinerariesResponse": {
            "type": "object",
            "properties": {
                "alternativeItineraries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/itineraries.ResponseItinerary"
                    }
                },
                "debugOutput": {
                    "type": "object",
                    "additionalProperties": true
                },
                "from": {
                    "$ref": "#/definitions/itineraries.ResponsePosition"
                },
                "requestParameters": {
                    "type": "object",
                    "additionalProperties": true
                },
                "safestItinerary": {
                    "$ref": "#/definitions/itineraries.ResponseItinerary"
                },
                "to": {
                    "$ref": "#/definitions/itineraries.ResponsePosition"
                }
            }
        },
        "itineraries.LegGeometry": {
            "type": "object",
            "properties": {
                "length": {
                    "type": "integer"
                },
                "points": {
                    "type": "string"
                }
            }
        },
        "itineraries.ResponseItinerary": {
            "type": "object",
            "properties": {
                "calculated_risk": {
                    "type": "number"
                },
                "duration": {
                    "type": "integer"
                },
                "endTime": {
                    "type": "string"
                },
                "legs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/itineraries.ResponseLeg"
                    }
                },
                "startTime": {
                    "type": "string"
                },
                "transfers": {
                    "type": "integer"
                }
            }
        },
        "itineraries.ResponseLeg": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "endTime": {
                    "type": "string"
                },
                "from": {
                    "$ref": "#/definitions/itineraries.ResponsePosition"
                },
                "intermediateStops": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/itineraries.ResponsePosition"
                    }
                },
                "legGeometry": {
                    "$ref": "#/definitions/itineraries.LegGeometry"
                },
                "mode": {
                    "type": "string"
                },
                "routeShortName": {
                    "type": "string"
                },
                "scheduledEndTime": {
                    "type": "string"
                },
                "scheduledStartTime": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                },
                "to": {
                    "$ref": "#/definitions/itineraries.ResponsePosition"
                }
            }
        },
        "itineraries.ResponsePosition": {
            "type": "object",
            "properties": {
                "arrival": {
                    "type": "string"
                },
                "departure": {
                    "type": "string"
                },
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "scheduledArrival": {
                    "type": "string"
                },
                "scheduledDeparture": {
                    "type": "string"
                },
                "stopId": {
                    "type": "string"
                }
            }
        },
...
        "statistics.Statistics": {
            "type": "object",
            "properties": {
                "numberOfReports": {
                    "type": "integer"
                }
            }
        },
        "utils.Coordinates": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
...
        "utils.InspectorRequest": {
            "type": "object",
            "properties": {
                "author": {
                    "description": "is always null or 98111116 (ASCII for \"BOT\") when getting data from the telegram bot",
                    "type": "integer"
                },
                "directionId": {
                    "type": "string"
                },
                "line": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "stationId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "utils.ResponseData": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "integer"
                },
                "direction": {
                    "$ref": "#/definitions/utils.Station"
                },
                "line": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "station": {
                    "$ref": "#/definitions/utils.Station"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "utils.Station": {
            "type": "object",
            "properties": {
                "coordinates": {
                    "$ref": "#/definitions/utils.Coordinates"
                },
                "id": {
                    "type": "string"
                },
                "lines": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "utils.TicketInspectorResponse": {
            "type": "object",
            "properties": {
                "direction": {
                    "$ref": "#/definitions/utils.Station"
                },
                "isHistoric": {
                    "type": "boolean"
                },
                "line": {
                    "description": "String is used so that it can easily be handled by the frontend",
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "station": {
                    "$ref": "#/definitions/utils.Station"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    }
}